+++ ikarus/finiteelements/mechanics/nonlinearelastic.hh
@@ -18,10 +18,9 @@
   #include <dune/localfefunctions/impl/standardLocalFunction.hh>
   #include <dune/localfefunctions/manifolds/realTuple.hh>
 
-  #include <ikarus/finiteelements/febases/powerbasisfe.hh>
+  #include <ikarus/finiteelements/febase.hh>
   #include <ikarus/finiteelements/fehelper.hh>
   #include <ikarus/finiteelements/ferequirements.hh>
-  #include <ikarus/finiteelements/fetraits.hh>
   #include <ikarus/finiteelements/mechanics/loads.hh>
   #include <ikarus/finiteelements/mechanics/materials/tags.hh>
   #include <ikarus/finiteelements/physicshelper.hh>
@@ -36,26 +35,29 @@ namespace Ikarus {
  *
  * \ingroup mechanics
  *
- * \tparam B The basis type for the finite element.
+ * \tparam BH The basis handler type for the finite element.
  * \tparam MAT The material type for the finite element.
  * \tparam FER The requirements for the finite element.
  * \tparam useEigenRef A boolean flag indicating whether to use Eigen references.
+ * \tparam useFlat A boolean indicating if the type of the underlying basis is of the flat or the untouched version.
  */
-template <typename B, typename MAT, typename FER = FERequirements<>, bool useEigenRef = false>
-class NonLinearElastic : public PowerBasisFE<B>,
-                         public Volume<NonLinearElastic<B, MAT, FER, useEigenRef>, FETraits<B, FER, useEigenRef>>,
-                         public Traction<NonLinearElastic<B, MAT, FER, useEigenRef>, FETraits<B, FER, useEigenRef>>
+template <typename BH, typename MAT, typename FER = FERequirements<>, bool useEigenRef = false, bool useFlat = true>
+class NonLinearElastic : public FEBase<BH, useFlat, FER, useEigenRef>,
+                         public Volume<NonLinearElastic<BH, MAT, FER, useEigenRef, useFlat>,
+                                       typename FEBase<BH, useFlat, FER, useEigenRef>::Traits>,
+                         public Traction<NonLinearElastic<BH, MAT, FER, useEigenRef, useFlat>,
+                                         typename FEBase<BH, useFlat, FER, useEigenRef>::Traits>
 {
 public:
-  using Traits            = FETraits<B, FER, useEigenRef>;
-  using Basis             = typename Traits::Basis;
+  using Base              = FEBase<BH, useFlat, FER, useEigenRef>;
+  using Traits            = typename Base::Traits;
+  using BasisHandler      = typename Traits::BasisHandler;
   using FlatBasis         = typename Traits::FlatBasis;
   using FERequirementType = typename Traits::FERequirementType;
   using LocalView         = typename Traits::LocalView;
   using Geometry          = typename Traits::Geometry;
   using GridView          = typename Traits::GridView;
   using Element           = typename Traits::Element;
-  using BasePowerFE       = PowerBasisFE<Basis>; // Handles globalIndices function
   using Material          = MAT;
   using VolumeType        = Volume<NonLinearElastic, Traits>;
   using TractionType      = Traction<NonLinearElastic, Traits>;
@@ -69,7 +71,7 @@ public:
    *
    * \tparam VolumeLoad The type for the volume load function.
    * \tparam NeumannBoundaryLoad The type for the Neumann boundary load function.
-   * \param globalBasis The global basis for the finite element.
+   * \param basisHandler The basis handler for the finite element.
    * \param element The element for which the finite element is constructed.
    * \param mat The material for the non-linear elastic element.
    * \param volumeLoad Volume load function (default is LoadDefault).
@@ -77,14 +79,13 @@ public:
    * \param neumannBoundaryLoad Neumann boundary load function (default is LoadDefault).
    */
   template <typename VolumeLoad = utils::LoadDefault, typename NeumannBoundaryLoad = utils::LoadDefault>
-  NonLinearElastic(const Basis& globalBasis, const typename LocalView::Element& element, const Material& mat,
+  NonLinearElastic(const BasisHandler& basisHandler, const typename LocalView::Element& element, const Material& mat,
                    VolumeLoad volumeLoad = {}, const BoundaryPatch<GridView>* neumannBoundary = nullptr,
                    NeumannBoundaryLoad neumannBoundaryLoad = {})
-      : BasePowerFE(globalBasis, element),
+      : Base(basisHandler, element),
         VolumeType(volumeLoad),
         TractionType(neumannBoundary, neumannBoundaryLoad),
         mat_{mat} {
-    this->localView().bind(element);
     auto& firstChild = this->localView().tree().child(0);
     const auto& fe   = firstChild.finiteElement();
     geo_             = std::make_shared<const Geometry>(this->localView().element().geometry());